requirements.txt
chatterbot==1.0.4
django-rosetta
django-crispy-forms
django_filter
pillow
django-jalali-date
